Update button location
